requirements.txt
Django==2.2
djangorestframework==3.9.4
psycopg2==2.7.3.2
pytz==2019.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2